phase3/requirements.txt
chromadb>=0.4.0
sentence-transformers>=2.2.0
python-dotenv>=1.0.0
pytest>=7.0
